dev: Change vcscore to __init__.py
Because python package recognize __init__.py
